requirements.txt
# appnope==0.1.0
# backcall==0.1.0
# beautifulsoup4==4.6.3
# bleach==2.1.4
# certifi==2018.8.24
# chardet==3.0.4
# Click==7.0
# cycler==0.10.0
# decorator==4.3.0
# defusedxml==0.5.0
# entrypoints==0.2.3
# Flask==1.0.2
# google==2.0.1
# google-cloud==0.34.0
# gtfs-realtime-bindings==0.0.5
# html5lib==1.0.1
# idna==2.7
ipykernel==5.0.0
# ipython==7.1.1
# ipython-genutils==0.2.0
# ipywidgets==7.4.2
# itsdangerous==1.1.0
# jedi==0.12.1
# Jinja2==2.10
# jsonschema==2.6.0
# jupyter==1.0.0
# jupyter-client==5.2.3
# jupyter-console==5.2.0
# jupyter-core==4.4.0
# kiwisolver==1.0.1
# MarkupSafe==1.0
matplotlib==3.3.3
PyOpenGL==3.1.5
pygame==2.0.1
# mistune==0.8.3
# nbconvert==5.4.0
# nbformat==4.4.0
# nltk==3.3
# notebook==5.7.0
numpy==1.15.3
# oauthlib==2.1.0
# pandas==0.23.3
# pandocfilters==1.4.2
# parso==0.3.1
# pexpect==4.6.0
# pickleshare==0.7.5
# prometheus-client==0.3.1
# prompt-toolkit==2.0.7
# protobuf==3.6.1
# protobuf-to-dict==0.1.0
# ptyprocess==0.6.0
# Pygments==2.2.0
# pyparsing==2.2.0
# PySocks==1.6.8
# python-dateutil==2.7.3
# pytz==2018.5
# pyzmq==17.1.2
# qtconsole==4.4.1
# requests==2.19.1
# requests-oauthlib==1.0.0
# Send2Trash==1.5.0
# simplegeneric==0.8.1
# six==1.11.0
# terminado==0.8.1
# testpath==0.4.1
# tornado==5.1.1
# traitlets==4.3.2
# tweepy==3.6.0
# urllib3==1.23
# virtualenv==16.0.0
# wcwidth==0.1.7
# webencodings==0.5.1
# Werkzeug==0.14.1
# widgetsnbextension==3.4.2
# xlrd==1.1.0